Compute seismic specificity score for each gene and cell type.

Compute seismic specificity score for each gene and cell type.

Usage

calc_specificity(
  sce,
  assay_name = "logcounts",
  ct_label_col = "idents",
  min_uniq_ct = 2,
  min_ct_size = 20,
  min_cells_gene_exp = 10,
  min_avg_exp_ct = 0.1
)

Arguments

sce

A SingleCellExperiment object. This object needs to include the assay_name specified and a ct_label_col column corresponding to cell type labels for the granularity of interest. Row names are used as the gene name identifiers.

assay_name

An assay in sce, default 'logcounts' (calculated using scran).

ct_label_col

A column name in colData of sce (information for the cells in sce) that provides labels for each cell, indicating membership in each cell type at the granularity of interest.

min_uniq_ct

The minimum number of unique cell types needed in the data to proceed with specificity score calculations. Default: 2.

min_ct_size

Filter for cell types to be included in the specificity score calculation based on the number of cells that are labeled as that cell type. Default: 20.

min_cells_gene_exp

Filter for genes to be included in the specificity score calculation based on the number of cells where the gene has non-zero expression. Default: 10.

min_avg_exp_ct

Filter for genes to be included in the specificity score calculation based on the mean of cell-type-averaged expressions (aka mean of mean cell type expression). Default: 0.1

Value

A dgeMatrix of specificity scores (genes as rows, cell types as columns).