Check magma object / file is usable and returns it as a data.table

Check magma object / file is usable and returns it as a data.table

Usage

load_magma_dt(magma, magma_gene_col, magma_z_col)

Arguments

magma

input magma object / file

magma_gene_col

character string corresponding to the column name in the MAGMA data containing gene identifiers

magma_z_col

character string corresponding to the column name in the MAGMA data containing z-scores

Value

magma file contents as a data.table