MAGMA genes.out output for rheumatoid arthritis

A data.frame of the genes.out output from MAGMA calculated using rheumatoid arthritis summary statistics.

Usage

ra_magma

Format

ra_magma

A data.frame with 17,460 rows and 9 columns:

GENE

entrez gene ID

CHR

the chromosome the gene is on

START

start boundary of the gene on the chromosome (including any provided window during annotation)

END

end boundary of the gene on the chromosome (including any provided window during annotation)

NSNPS

number of SNPS annotated to gene (not excluded based on internal SNP QC)

NPARAM

number of relevant parameters used in the MAGMA model

N

sample size used in analyzing the gene

ZSTAT

z-value of gene based on permuted p-value (what we also term MAGMA z-score)

P

gene p-value

Source

https://www.nature.com/articles/nature12873