Translate gene ids of a given dgeMatrix of seismic specificity scores from a given organism (e.g., mouse) to another (e.g., human) based on orthology.

Currently supported organisms: hsa, mmu (human, mouse, respectively) Currently supported IDs: symbol, ensembl, entrez

Usage

translate_gene_ids(
  sscore,
  from = "mmu_ensembl",
  to = "hsa_entrez",
  multi_mapping = "mean"
)

Arguments

sscore

dgeMatrix of specificity scores (direct output of calc_specificity())

from

The organism and ID type of current identifiers, separated by an underscore (default: mmu_ensembl).

to

The organism and ID type of target identifiers, separated by an underscore (default: hsa_entrez).

multi_mapping

Method to translate metrics when one source ID has several target ID matches. Options: mean (default), sum.

Value

A dgeMatrix that has row names using target identifiers, with

Details

Unmatched rows are dropped from the matrix.